move fetch scene ids and json to separate utils files
